chore(internal.snmp): Move common SNMP parts to plugins/common
