+++ plugins/common/snmp/testdata/gosmi/foo
