+++ plugins/common/snmp/testdata/gosmi/ifPhysAddress
