+++ plugins/common/snmp/testdata/gosmi/ifPhysAddressImports
