+++ plugins/common/snmp/testdata/gosmi/server
