+++ plugins/common/snmp/testdata/gosmi/tableBuild
