+++ plugins/common/snmp/testdata/gosmi/tableMib
