+++ plugins/common/snmp/testdata/loadMibsFromPath/root/symlink
