+++ plugins/common/snmp/translator_gosmi.go
