+++ plugins/common/snmp/translator_gosmi_test.go
