+++ plugins/common/snmp/translator_netsnmp_mocks_generate.go
