+++ plugins/common/snmp/translator_netsnmp_mocks_test.go
