+++ plugins/common/snmp/translator_netsnmp_test.go
