+++ plugins/inputs/snmp_trap/gosmi.go
@@ -2,7 +2,7 @@ package snmp_trap
 
 import (
 	"github.com/influxdata/telegraf"
-	"github.com/influxdata/telegraf/internal/snmp"
+	"github.com/influxdata/telegraf/plugins/common/snmp"
 )
 
 type gosmiTranslator struct {
